# Update `parse_definitions` tool to cache regex compilation in generated file