… lib/src/magic/generated.rs