Thank you for this great crate! 

This PR updates the `parse_definitions` tool introducing a new `RuleRegexes` type which serves to gather a deduplicated set of regexes per mimetype. When rendered to the generated file, these regexes are wrapped in a `LazyLock` which removes the need for recompilation. Regex matching in the `check` is done against these static/lazily instantiated versions instead of recompiling for every call.

This is motivated by a use case where I'm importing 500k files and running a mime type detector over the top of them before putting them in a SQLite database. With the existing release version of the library it spent a great deal simply compiling regexes. With the change I've made here I got a ~10x speed up. 

An interesting question might be trading off one shot performance versus caching the compilation, since there is presumably some overhead with the `LazyLock`. I ran the existing `cargo bench` suite and it showed no major differences, or some improvement.

Before:
```
test tika-magic::from_u8::application_zip  ... bench:       1,624 ns/iter (+/- 15)
test tika-magic::from_u8::image_gif        ... bench:          18 ns/iter (+/- 0)
test tika-magic::from_u8::image_png        ... bench:          10 ns/iter (+/- 0)
test tika-magic::from_u8::text_plain       ... bench:   5,232,508 ns/iter (+/- 71,231)
test tika-magic::match_u8::application_zip ... bench:         469 ns/iter (+/- 3)
test tika-magic::match_u8::image_gif       ... bench:          13 ns/iter (+/- 0)
test tika-magic::match_u8::image_png       ... bench:          13 ns/iter (+/- 0)
test tika-magic::from_u8::text_plain       ... bench:       8,743 ns/iter (+/- 246)
```

After:
```
test tika-magic::from_u8::application_zip ... bench:       1,566 ns/iter (+/- 64)
test tika-magic::from_u8::image_gif       ... bench:          17 ns/iter (+/- 1)
test tika-magic::from_u8::image_png       ... bench:           9 ns/iter (+/- 0)
test tika-magic::from_u8::text_plain      ... bench:      30,243 ns/iter (+/- 471)
test tika-magic::from_u8::application_zip ... bench:         461 ns/iter (+/- 16)
test tika-magic::from_u8::image_gif       ... bench:          13 ns/iter (+/- 0)
test tika-magic::from_u8::image_png       ... bench:          12 ns/iter (+/- 1)
test tika-magic::from_u8::text_plain      ... bench:          25 ns/iter (+/- 0)
```

Note! I have not updated the generated file in the repo because I wasn't 100% confident on how you'd like this to be done. Happy to do this with guidance!