@withzombies 2025-06-22
Hey! Thank you for the PR. I was complaining about rust regex initialization when I wrote this because it bothered me too! I'm happy to merge this.

On generating the rules, I just run these few commands. Nothing fancy but also not really plumbed in to be done automatically.
```sh
cargo run --bin parse_definitions definitions/mimetypes.xml
mv generated.rs lib/src/magic/generated.rs
cargo fmt --all
```